Fix `UnsafeAtomic`'s `Sendable` conformance
In Swift 5.5, `UnsafeAtomic` has gained an implicit, unconditional `Sendable` conformance, because it is a `@frozen public struct` whose sole stored property (an `UnsafeMutablePointer` is also `Sendable`. This is not correct -- we only want `UnsafeAtomic` to conform to `Sendable` if its `Value` does.

Explicitly declare the conformance on 5.5 to disable the bogus inference.