Install swiftmodules with full module triple
Clean up how the CMake build installs swiftmodules using the full module
triple instead of just the architecture. This change uses the compiler
to compute the expected module triple instead of attempting to compute
it in CMake script. This is more maintainable and robust when adding new
platforms and architectures.